refactor(rskim-search): split types.rs into focused sub-modules
types.rs was 577 lines, exceeding the 400-line project convention.
Split into four focused modules under types/:

- file_table.rs  — FileId, FileTable, custom serde impls, normalize(),
                   64-bit compile_error gate
- query.rs       — SearchQuery (#[must_use]), TemporalFlags, SearchField
- result.rs      — SearchResult, MatchSpan, LineRange, IndexStats
- error.rs       — SearchError, Result alias

mod.rs barrel re-exports all types so lib.rs pub use is unchanged.
Inline unit tests moved into their respective sub-modules.

All 91 tests pass; clippy --workspace -D warnings clean.
